refactor(media): move comments into video viewer
