+++ frontend/src/features/video-viewer/comments/hooks/usePlayerReady.js
