+++ frontend/src/features/video-viewer/comments/utils/avatar.js
